Refactor agent daemon: consolidate into daemon package and extract phase helpers
- Move nodeupdate package to cmd/agent/internal/daemon, rename Execute to
  updateNode (unexported). The daemon package now owns the full watch loop,
  reconciliation, kube client, scheme, drift detection, and node update logic.

- Slim cmd/daemon.go to just cobra wiring that calls daemon.Run(ctx, log).
  Delete cmd/scheme.go (moved to daemon package).

- Remove Config field from NodeStart goal state. PersistAppliedConfig and
  StartNode now take *provision.AgentConfig as an explicit parameter instead
  of smuggling it through the goal state struct.

- Extract WaitForKubelet into phases/nodestart/wait_kubelet.go.

- Extract PersistAppliedConfig into phases/nodestart/persist_config.go.

- Add reset.RemoveAppliedConfig task and reset.CleanupMachine composite
  that combines RemoveNSpawnConfig + RemoveMachine + RemoveAppliedConfig.
  Replace inline os.Remove in node update with the composite task.
