+++ cmd/agent/internal/goalstates/constants.go
@@ -40,6 +40,7 @@ func AlternateMachine(current string) string {
 	if current == NSpawnMachineKube1 {
 		return NSpawnMachineKube2
 	}
+
 	return NSpawnMachineKube1
 }
